refactor(openrouter): move request augmentation behind core seam
Introduce an `OpenResponsesRequestExtension` hook on core's Open Responses
driver and move all OpenRouter request-body augmentation out of core into
the everruns-openrouter crate.

- core: the driver serializes the vendor-neutral request to a JSON value,
  then lets an optional provider extension decorate it before sending.
  Removes the `models`/`route`/`provider`/`session_id`/`plugins` fields and
  every `provider_type == OpenRouter` branch (plus `plugins_to_wire`) from
  the shared driver, leaving it provider-neutral.
- openrouter: `OpenRouterRequestExtension` carries the routing validation,
  preset/capacity resolution, plugins serialization, and session_id
  forwarding; the driver attaches it. Wire-level tests move to the crate.

Routing data types stay in core, so model_router keeps building them with
no dependency cycle. Behavior is unchanged; the request body sent to
OpenRouter is identical.
