+++ crates/openrouter/src/lib.rs
@@ -21,9 +21,11 @@
 //! ```
 
 mod driver;
+mod request_ext;
 mod types;
 
 pub use driver::{OpenRouterChatDriver, register_driver};
+pub use request_ext::OpenRouterRequestExtension;
 pub use types::{
     OpenRouterArchitecture, OpenRouterModelInfo, OpenRouterModelsResponse, OpenRouterPricing,
     OpenRouterTopProvider,
